requirements.txt
flask==3.0.0
pandas==2.2.2
flask_cors==4.0.1
matplotlib==3.9.1
mysql-connector-python==9.0.0
urllib3 >= 1.26.11
requests >= 2.25.1
pandas >= 2.2.2
bs4 >= 0.0.1
regex >= 2023.10.3
selenium >= 4.14.0
newspaper3k >= 0.2.8
openai >= 1.35.13
googlemaps >= 4.10.0
dateparser >= 1.2.0
spacy >= 3.7.6
python-dotenv >= 1.0.1
gunicorn == 23.0.0